+++ packages/gov-ui-kit/src/modules/components/odsModulesProvider/odsModulesProvider.tsx
@@ -1,7 +1,7 @@
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { createContext, useContext, useMemo, type ReactNode } from 'react';
 import { createClient, http } from 'viem';
-import { WagmiProvider, createConfig, type Config } from 'wagmi';
+import { WagmiProvider, createConfig, type Config, type State } from 'wagmi';
 import { arbitrum, arbitrumSepolia, base, baseSepolia, mainnet, polygon, polygonAmoy, sepolia } from 'wagmi/chains';
 import { OdsCoreProvider, type IOdsCoreProviderProps } from '../../../core';
 import { modulesCopy, type ModulesCopy } from '../../assets';
@@ -35,6 +35,10 @@ export interface IOdsModulesProviderProps {
      */
     wagmiConfig?: Config;
     /**
+     * Optional initial state for Wagmi provider.
+     */
+    wagmiInitialState?: State;
+    /**
      * React-query configurations to be forwarded to the QueryClientProvider, uses the defaults configurations from
      * react-query when not specified (see https://tanstack.com/query/latest/docs/framework/react/guides/important-defaults).
      * @default defaultQueryClient
@@ -65,6 +69,7 @@ export const OdsModulesProvider: React.FC<IOdsModulesProviderProps> = (props) =>
     const {
         queryClient = defaultQueryClient,
         wagmiConfig = defaultWagmiConfig,
+        wagmiInitialState,
         coreProviderValues,
         children,
         values,
@@ -79,7 +84,7 @@ export const OdsModulesProvider: React.FC<IOdsModulesProviderProps> = (props) =>
 
     return (
         <odsModulesContext.Provider value={contextValues}>
-            <WagmiProvider config={wagmiConfig}>
+            <WagmiProvider config={wagmiConfig} initialState={wagmiInitialState}>
                 <QueryClientProvider client={queryClient}>
                     <OdsCoreProvider values={coreProviderValues}>{children}</OdsCoreProvider>
                 </QueryClientProvider>